[741] First batch of fixes
