+++ src/pages/routines/index.ts
@@ -1,4 +1,4 @@
 export * from "./DeleteRoutinePage";
-export * from "./EditRoutinePage";
+export * from "./EditRoutinePage/EditRoutinePage";
 export * from "./NewRoutinePage";
 export * from "./RoutinesPage";
